Simple version of DirectToDbApplier
This is a really simple implementation that won't work in any proper
scenario.  The integration tests pass, but that's becuase they're a bit
too simple (all the scripts only have one statement in them)!

Still TODO:
1. The script file needs parsing to separate out multiple statements
2. The exception handling is messy and throwing a RuntimeException is
not a great solution (I'd love to use spring-jdbc, but thats almost 2MB's
worth of dependency ...)
3. The role of QueryExecuter is confused - it was originally invented
to make testing DatabaseSchemaVersionManager easier but now doesn't
really seem to be adding a whole lot of value.