Test that confirms change scripts with mutliple statements work
I expected this to fail!  The ant sql task goes to a lot of trouble
to separate the scripts it is told to execute based on delimiters.

But, with hsqldb at least, I can just submit a bit of sql with multiple
statements in it and they all work.

Now to do some experimentation to see if this is an hsqldb "enhancement"
or if this works with most databases....